// apps/react-client/src/types/assessment.ts
export interface ResourceInfoDto {
  id: number;
  title?: string;
  type?: string;
  url?: string;
}

export interface AssessmentResourceDto {
  id: number;
  resource: ResourceInfoDto;
}
// ...
export interface AssessmentListItemDto {
  id: number;
  title: string;
  instruction: string;
  dueDate: Date;
  startDate: Date;
  maxScore: number;
  session: number;
  isPublic: boolean;
  submissionType: SubmissionType;
  allowLate: boolean;
  penaltyCriteria?: string;
  aiEvaluationEnable: boolean;
  aiModelSelectionMode: AIModelSelectionMode;
  allowedSubmissionMethod: SubmissionMethod;
  created_at?: Date;
  updated_at?: Date;
  resources?: AssessmentResourceDto[];
}